% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-wallets.R
\name{fund.wallets}
\alias{fund.wallets}
\title{Fund wallets}
\usage{
fund.wallets(x, amount, fee.priority = "priority")
}
\arguments{
\item{x}{x}

\item{amount}{amount}

\item{fee.priority}{fee.priority}
}
\value{
1
}
\description{
Fund wallets
}
\examples{
1
}
