% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-wallets.R
\name{gen.wallets}
\alias{gen.wallets}
\title{Generate wallets}
\usage{
gen.wallets(
  monerod.rpc.port,
  n.wallets,
  wallets.data.file = "spam_wallets_data.rds",
  print.json = TRUE,
  cleanup.process.monero_wallet_rpc = FALSE,
  monero.wallet.rpc.log.level = 0
)
}
\arguments{
\item{monerod.rpc.port}{monerod.rpc.port}

\item{n.wallets}{n.wallets}

\item{wallets.data.file}{wallets.data.file}

\item{print.json}{print.json}

\item{cleanup.process.monero_wallet_rpc}{cleanup.process.monero_wallet_rpc}

\item{monero.wallet.rpc.log.level}{monero.wallet.rpc.log.level}
}
\value{
1
}
\description{
Generate wallets
}
\examples{
1
}
