% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-leaves.R
\name{prep.leaves.wallets}
\alias{prep.leaves.wallets}
\title{Prepare Wallet Account Tree Leaves}
\usage{
prep.leaves.wallets(
  wallets,
  threads = NA,
  n.outputs = 15,
  n.tree.levels = 4,
  fee.priority = c(3, 3, 3, 0),
  wallets.data.file = "spam_wallets_data.rds"
)
}
\arguments{
\item{wallets}{wallets}

\item{threads}{threads}

\item{n.outputs}{n.outputs}

\item{n.tree.levels}{n.tree.levels}

\item{fee.priority}{fee.priority}

\item{wallets.data.file}{wallets.data.file}
}
\value{
1
}
\description{
Prepare Wallet Account Tree Leaves
}
\examples{
1
}
