% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{revive.wallets}
\alias{revive.wallets}
\title{Revive wallets processes}
\usage{
revive.wallets(
  monerod.rpc.port,
  wallets.data.file = "spam_wallets_data.rds",
  restart.alive.processes = FALSE,
  print.json = FALSE,
  cleanup.process.monero_wallet_rpc = FALSE
)
}
\arguments{
\item{monerod.rpc.port}{monerod.rpc.port}

\item{wallets.data.file}{wallets.data.file}

\item{restart.alive.processes}{restart.alive.processes}

\item{print.json}{print.json}

\item{cleanup.process.monero_wallet_rpc}{cleanup.process.monero_wallet_rpc}
}
\value{
1
}
\description{
Revive wallets processes
}
\examples{
1
}
