% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{spam.1in.2out.wallets}
\alias{spam.1in.2out.wallets}
\title{Spam 1in/2out transactions}
\usage{
spam.1in.2out.wallets(
  wallets,
  threads = NA,
  fee.priority = 1,
  delay.between.each.tx = 0
)
}
\arguments{
\item{wallets}{wallets}

\item{threads}{threads}

\item{fee.priority}{fee.priority}

\item{delay.between.each.tx}{delay.between.each.tx}
}
\value{
1
}
\description{
Spam 1in/2out transactions
}
\examples{
1
}
