% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{wallets.status}
\alias{wallets.status}
\title{Check wallets status}
\usage{
wallets.status(wallets, get_balance = FALSE, log.lines = 2)
}
\arguments{
\item{wallets}{wallets}

\item{get_balance}{get_balance}

\item{log.lines}{log.lines}
}
\value{
1
}
\description{
Check wallets status
}
\examples{
1
}
