% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmrspam-package.R, R/xmrspammer-package.R
\docType{package}
\name{xmrspammer-package}
\alias{xmrspammer-package}
\title{xmrspammer: Create high volume of spam Monero transactions}
\description{
Functions for managing Monero wallets to create many transactions in a short period of time. Useful for Monero "stressnet" testing.

Functions for managing Monero wallets to create many transactions in a short period of time. Useful for Monero "stressnet" testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rucknium.github.io/xmrspammer/}
}


Useful links:
\itemize{
  \item \url{https://rucknium.github.io/xmrspammer/}
}

}
\author{
\strong{Maintainer}:  Rucknium \email{Rucknium@protonmail.com} (\href{https://orcid.org/0000-0001-5999-8950}{ORCID}) [copyright holder]

}
\keyword{internal}
